% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adams_bt_03}
\alias{adams_bt_03}
\title{Adams Lake Bull Trout Population Parameters}
\format{An object of class \code{\link{ypr_population}}.}
\usage{
adams_bt_03
}
\description{
A list of the estimates of the life-history parameters for
Bull Trout in Adams Lake, BC, adapted from Bison et al (2003).
}
\examples{
ypr_plot(adams_bt_03)
}
\references{
Bison, R., O’Brien, D., and Martell, S.J.D. 2003. An Analysis of Sustainable Fishing Options for Adams Lake Bull Trout Using Life History and Telemetry Data. BC Ministry of Water Land and Air Protection, Kamloops, B.C.
}
\keyword{datasets}
