% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-ecotypes.R
\name{as_ypr_ecotypes}
\alias{as_ypr_ecotypes}
\title{Ecotype Parameters}
\usage{
as_ypr_ecotypes(populations, weights, names = NULL)
}
\arguments{
\item{populations}{A list with all parameters for each ecotype in the
population}

\item{weights}{A numeric vector giving the weighting for how much it
contributes to the overall population}

\item{names}{A character vector providing names for each ecotype. Default is
NULL.}
}
\description{
Creates an object of class ecotype.
}
