% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-populations.R
\name{as_ypr_populations}
\alias{as_ypr_populations}
\alias{as_ypr_populations.data.frame}
\title{Coerce to a Populations Parameter Object}
\usage{
as_ypr_populations(x, ...)

\method{as_ypr_populations}{data.frame}(x, ...)
}
\arguments{
\item{x}{The object to coerce.}

\item{...}{Unused parameters.}
}
\description{
Coerce to a Populations Parameter Object
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Coerce data frame to ypr_populations
}}

\examples{
as_ypr_populations(as.data.frame(ypr_populations(Rk = c(3, 4))))
}
\seealso{
Other populations: 
\code{\link{chilliwack_bt_05}},
\code{\link{update.ypr_populations}()},
\code{\link{ypr_plot_yield}()},
\code{\link{ypr_population_names}()},
\code{\link{ypr_populations_expand}()},
\code{\link{ypr_populations}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}
}
\concept{populations}
