% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_ecotypes}
\alias{check_ecotypes}
\title{Check Ecotypes}
\usage{
check_ecotypes(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
An informative error if the test fails or an invisible copy of x.
}
\description{
Checks if an ypr_ecotypes object with valid parameter values.
}
\examples{
check_ecotypes(ypr_ecotypes())
}
\seealso{
Other check: 
\code{\link{check_populations}()},
\code{\link{check_population}()}
}
\concept{check}
