% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_population}
\alias{check_population}
\title{Check Population}
\usage{
check_population(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
An informative error if the test fails or an invisible copy of x.
}
\description{
Checks if an ypr_population object with valid parameter values.
}
\examples{
check_population(ypr_population())
}
\seealso{
Other check: 
\code{\link{check_ecotypes}()},
\code{\link{check_populations}()}
}
\concept{check}
