% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chilliwack_bt_05}
\alias{chilliwack_bt_05}
\title{Chilliwack Lake Bull Trout Population Parameters}
\format{An object of class \code{\link{ypr_population}}.}
\usage{
chilliwack_bt_05
}
\description{
A list of the estimates of the life-history parameters for
Bull Trout in Chilliwack Lake, BC, adapted from Taylor et al (2005).
}
\examples{
ypr_plot(chilliwack_bt_05)
ypr_plot(ypr_population_update(chilliwack_bt_05, Rk = 2.5, Hm = 0.05, rho = 1))
}
\references{
Taylor, J.L. 2005. Sustainability of the Chilliwack Lake Char Fishery. Ministry of Water, Land and Air Protection, Surrey, B.C.
}
\keyword{datasets}
