% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.ypr_population}
\alias{is.ypr_population}
\alias{is_ypr_population}
\alias{is.ypr_populations}
\alias{is_ypr_populations}
\alias{is.ypr_ecotypes}
\alias{is_ypr_ecotypes}
\title{Tests if is a Population, Populations or Ecotypes}
\usage{
is.ypr_population(x)

is_ypr_population(x)

is.ypr_populations(x)

is_ypr_populations(x)

is.ypr_ecotypes(x)

is_ypr_ecotypes(x)
}
\arguments{
\item{x}{The object to test.}
}
\description{
Tests if is a Population, Populations or Ecotypes
}
\section{Functions}{
\itemize{
\item \code{is_ypr_population()}: Test if is a Population

\item \code{is.ypr_populations()}: Test if is a Populations

\item \code{is_ypr_populations()}: Test if is a Populations

\item \code{is.ypr_ecotypes()}: Test if is an Ecotypes

\item \code{is_ypr_ecotypes()}: Test if is an Ecotypes

}}
\examples{
is.ypr_population(ypr_population())
is_ypr_population(ypr_population())
is.ypr_populations(ypr_populations())
is_ypr_population(ypr_populations())
is.ypr_ecotypes(ypr_ecotypes())
is_ypr_ecotypes(ypr_ecotypes())
}
