% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kootenay_rb}
\alias{kootenay_rb}
\title{Kootenay Lake Rainbow Trout Population Parameters}
\format{
An object of class \code{\link[=ypr_population]{ypr_population()}}.
}
\usage{
kootenay_rb
}
\description{
The population parameters for
Rainbow Trout in Kootenay Lake.
}
\details{
The estimates are liable to change and should not be used for management.
}
\examples{
kootenay_rb
ypr_plot_yield(kootenay_rb)
}
\references{
Thorley, J.L., and Andrusak, G.F. 2017. The fishing and natural mortality of large, piscivorous Bull Trout and Rainbow Trout in Kootenay Lake, British Columbia (2008–2013). PeerJ 5: e2874. doi:10.7717/peerj.2874.
}
\keyword{datasets}
