% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kootenay_rb_13}
\alias{kootenay_rb_13}
\title{Kootenay Lake Rainbow Trout Population Parameters (2013)}
\format{An object of class \code{\link{ypr_population}}.}
\usage{
kootenay_rb_13
}
\description{
A list of the estimates of the life-history parameters for
Rainbow Trout in Kootenay Lake, BC, adapted from Andrusak and Thorley (2013).
The estimates should not be used for management.
}
\examples{
ypr_plot(kootenay_rb_13)
ypr_plot(kootenay_rb_13, Ly = 80)
}
\references{
Andrusak, G.F., and Thorley, J.L. 2013. Kootenay Lake Exploitation Study: Fishing and Natural Mortality of Large Rainbow Trout and Bull Trout: 2013 Annual Report. A Poisson Consulting Ltd. and Redfish Consulting Ltd. Report, Habitat Conservation Trust Foundation, Victoria, BC.
}
\keyword{datasets}
