% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quesnel_bt}
\alias{quesnel_bt}
\title{Quesnel Lake Bull Trout Population Parameters}
\format{An object of class \code{\link[=ypr_population]{ypr_population()}}.}
\usage{
quesnel_bt
}
\description{
The population parameters for
Bull Trout in Quesnel Lake, BC.
}
\details{
The estimates are liable to change and should not be used for management.
}
\examples{
quesnel_bt
ypr_plot_yield(quesnel_bt)
}
\keyword{datasets}
