% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quesnel_lt}
\alias{quesnel_lt}
\title{Quesnel Rainbow Trout Population Parameters}
\format{An object of class \code{\link{ypr_population}}.}
\usage{
quesnel_lt
}
\description{
A list of the current best estimates of the life-history parameters for
Rainbow Trout in Quesnel Lake, BC.
The estimates are liable to change.
}
\examples{
ypr_plot(quesnel_lt)
}
\keyword{datasets}
