% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quesnel_rb}
\alias{quesnel_rb}
\title{Quesnel Lake Rainbow Trout Population Parameters}
\format{
An object of class \code{\link[=ypr_population]{ypr_population()}}.
}
\usage{
quesnel_rb
}
\description{
The population parameters for
Rainbow Trout in Quesnel Lake, BC.
}
\details{
The estimates are liable to change and should not be used for management.
}
\examples{
quesnel_rb
ypr_plot_yield(quesnel_rb)
}
\keyword{datasets}
