% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr-package.R
\docType{package}
\name{ypr-package}
\alias{ypr}
\alias{ypr-package}
\title{ypr: Yield Per Recruit}
\description{
An implementation of equilibrium-based
yield per recruit methods.
The yield can be based on the number of fish caught (or harvested) or
biomass caught (or harvested) for
all fish or just large (trophy) individuals.
The yield (and optimal capture rate) can be calculated with
lower and upper limits that represent the uncertainty in
one or more population parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/ypr}
  \item Report bugs at \url{https://github.com/poissonconsulting/ypr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
