% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr.R
\name{ypr}
\alias{ypr}
\title{Yield-Per-Recruit}
\usage{
ypr(tmax = 20L, k = 0.5, Linf = 1000, t0 = 0, a = 1e-05, b = 3)
}
\arguments{
\item{tmax}{The maximum age.}

\item{k}{The growth coefficient.}

\item{Linf}{The mean maximum length.}

\item{t0}{The age at length 0.}

\item{a}{The condition factor.}

\item{b}{The scaling exponent.}
}
\description{
Yield-Per-Recruit
}
\examples{
ypr()
}
