% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_detabulate_parameters}
\alias{ypr_detabulate_parameters}
\title{Detabulate Parameters}
\usage{
ypr_detabulate_parameters(x, chk = TRUE)
}
\arguments{
\item{x}{A data frame with columns Parameter and Value specifying
one or more parameters and their values.}

\item{chk}{A flag specifying whether to check the other parameters.}
}
\value{
An object of class \code{\link[=ypr_population]{ypr_population()}}
}
\description{
Detabulate Parameters
}
\examples{
ypr_detabulate_parameters(ypr_tabulate_parameters(ypr_population()))
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}}
}
