% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_ecotypes}
\alias{ypr_ecotypes}
\title{Ecotypes}
\usage{
ypr_ecotypes(...)
}
\arguments{
\item{...}{One or more of the arguments from \code{ypr_population()}.}
}
\value{
A list of \code{\link{ypr_population}} objects each representing
a unique ecotype
with identical recruitment (\code{tR} \code{BH} and \code{Rk}) and
fishery (\code{Llo}, \code{Lup}, \code{NC}, \code{pi} and \code{q}) parameters.
The relative proportion of the recruits belonging to each ecotype is indicated
by the relative size of the \code{Rmax} parameters.

An object of class \code{ypr_ecotypes}.
}
\description{
Ecotypes
}
\examples{
ypr_ecotypes(Linf = c(50, 100))
}
