% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecundity.R
\name{ypr_fecundity}
\alias{ypr_fecundity}
\title{Calculate Fecundity}
\usage{
ypr_fecundity(L, Lm = 50, fa = 1e-05, fb = 3)
}
\arguments{
\item{L}{The length.}

\item{Lm}{The length at maturity.}

\item{fa}{The fecundity factor.}

\item{fb}{The fecundity scaling exponent.}
}
\value{
The expected weight for an individual of length L.
}
\description{
Calculates the expected fecundity for an individual of length L based on
an allometric equation.
}
\examples{
plot(0:100, ypr_fecundity(0:100), type = "l")
}
