% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{ypr_get_par}
\alias{ypr_get_par}
\title{Get Parameter Value}
\usage{
ypr_get_par(object, par = "pi")
}
\arguments{
\item{object}{A ypr object.}

\item{par}{A string of the parameter.}
}
\value{
A numeric or integer scalar or vector of the parameter value.
}
\description{
Get Parameter Value
}
\examples{
ypr_get_par(ypr_population())
}
