% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ypr_ggplot2}
\alias{ypr_ggplot2}
\title{Bivariate Plot of Schedule Terms}
\usage{
ypr_ggplot2(population, x = "Age", y = "Length")
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}

\item{x}{A string of the term on the x-axis.}

\item{y}{A string of the term on the y-axis.}
}
\value{
A ggplot2 object.
}
\description{
Produces a ggplot2 bivariate line plot of two schedule terms.
}
\examples{
ypr_ggplot2(ypr_population())
}
