% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_import_parameters}
\alias{ypr_import_parameters}
\title{Import Parameters}
\usage{
ypr_import_parameters(x)
}
\arguments{
\item{x}{A data frame with columns Parameter and Value specifying
one or more parameters and their values.}
}
\value{
An object of class \code{\link{ypr_population}}
}
\description{
Import Parameters
}
\examples{
ypr_import_parameters(ypr_tabulate_parameters(ypr_population()))
}
\seealso{
\code{\link{ypr_population}}
}
