% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypr_inst2inter}
\alias{ypr_inst2inter}
\title{Instantaneous to Interval Mortality}
\usage{
ypr_inst2inter(x)
}
\arguments{
\item{x}{A numeric vector of instantaneous mortality rates.}
}
\value{
A numeric vector of corresponding interval mortality rates.
}
\description{
Instantaneous to Interval Mortality
}
\examples{
ypr_inst2inter(c(0,0.2,3))
}
