% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypr_instant2interval}
\alias{ypr_instant2interval}
\title{Instantaneous to Interval Mortality}
\usage{
ypr_instant2interval(x)
}
\arguments{
\item{x}{A vector of instantaneous mortality rates.}
}
\value{
A vector of corresponding interval mortality rates.
}
\description{
Instantaneous to Interval Mortality
}
\examples{
ypr_instant2interval(c(0,0.2,3))
}
