% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypr_inter2inst}
\alias{ypr_inter2inst}
\title{Interval to Instantaneous Mortality}
\usage{
ypr_inter2inst(x)
}
\arguments{
\item{x}{A numeric vector of interval mortality rates.}
}
\value{
A numeric vector of corresponding instantaneous mortality rates.
}
\description{
Interval to Instantaneous Mortality
}
\examples{
ypr_inter2inst(c(0,0.5,0.99,1))
}
