% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypr_interval2instant}
\alias{ypr_interval2instant}
\title{Interval to Instantaneous Mortality}
\usage{
ypr_interval2instant(x)
}
\arguments{
\item{x}{A vector of interval mortality rates.}
}
\value{
A vector of corresponding instantaneous mortality rates.
}
\description{
Interval to Instantaneous Mortality
}
\examples{
ypr_interval2instant(c(0,0.5,0.99,1))
}
