% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{ypr_length}
\alias{ypr_length}
\title{Calculate Length}
\usage{
ypr_length(t, k = 1/2, Linf = 100, t0 = 0)
}
\arguments{
\item{t}{The age.}

\item{k}{The growth coefficient.}

\item{Linf}{The mean maximum length.}

\item{t0}{The age at length 0.}
}
\value{
The expected length for an individual of age t.
}
\description{
Calculates the expected length for an individual of age t based on the
Von Bertalanffy growth equation.
}
\examples{
plot(0:10, ypr_length(0:10), type = "l")
}
