% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypr_length_at_age}
\alias{ypr_length_at_age}
\title{Length At Age}
\usage{
ypr_length_at_age(population, age)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}

\item{age}{A numeric vector of the age (yr).}
}
\value{
A double vector of the lengths.
}
\description{
Length At Age
}
\examples{
ypr_length_at_age(ypr_population(), seq(0, 5, by = 0.5))
}
