% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names-deprec.R, R/names.R
\name{ypr_population_names}
\alias{ypr_population_names}
\alias{ypr_names}
\title{Population or Ecotype Names}
\usage{
ypr_population_names(populations)

ypr_names(x)
}
\arguments{
\item{populations}{An object of class \code{\link[=ypr_populations]{ypr_populations()}}.}

\item{x}{An object of class ypr_population, ypr_populations or ypr_ecotypes.}
}
\value{
A character vector of the unique parameter based names.
}
\description{
Generates set of unique names based on differences in parameter values.
}
\details{
Parameter RPR is ignored because it is irrelevant to population(s) and does not
distinguish between ecotypes.
}
\section{Functions}{
\itemize{
\item \code{ypr_population_names}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}}

\examples{
ypr_names(ypr_populations(Rk = c(2.5, 3, 2.5)))
}
