% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{ypr_optimize}
\alias{ypr_optimize}
\alias{ypr_optimise}
\title{Optimize Capture}
\usage{
ypr_optimize(population = ypr_population(), Ly = 0, harvest = TRUE,
  biomass = TRUE, check = FALSE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{Ly}{The minimum length fish to consider.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{check}{A flag indicating whether to check the arguments.}
}
\description{
Calculates the optimal yield for the population based on alternative
capture rates.
}
\examples{
ypr_optimize()
}
