% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-uncertainty.R
\name{ypr_optimize_uncertainty}
\alias{ypr_optimize_uncertainty}
\alias{ypr_optimise_uncertainty}
\title{Optimize Capture with Uncertainty}
\usage{
ypr_optimize_uncertainty(population, population2, Ly = 0, harvest = TRUE,
  biomass = TRUE, n = 10^4, level = 0.95,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{population2}{A second list of population life-history parameters.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{n}{A numeric of the number of iterations.}

\item{level}{A probability of the interval.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously.
The option is initialized from environment variable MC_CORES if set.
Must be at least one, and parallelization requires at least two cores.}
}
\description{
Finds the capture rate that maximises the yield for a given population with uncertainty.
The uncertainty in each population parameter is assumed to be a uniform distribution between the value in population and the value in population2.
The parameters are assumed to be uncorrelated and there is no requirement
for the value in population to be less than the value in population2
The parameters \code{tmax} and \code{Rk} must not vary.
}
\examples{
population <- ypr_population()
population$Rk <- 10
ypr_optimize_uncertainty(population, ypr_population(), n = 10)
}
\seealso{
\code{\link{ypr_population}}, \code{\link{ypr_yields}} and
\code{\link{ypr_optimize}}
}
