% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ypr_plot_sr}
\alias{ypr_plot_sr}
\title{Plot Stock-Recruitment Curve}
\usage{
ypr_plot_sr(population)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}
}
\value{
A ggplot2 object.
}
\description{
Plot Stock-Recruitment Curve
}
\examples{
ypr_plot_sr(ypr_population(Rk = 10))
ypr_plot_sr(ypr_population(Rk = 10, BH = 0L))
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_sr}}
}
