% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{ypr_population_names}
\alias{ypr_population_names}
\title{Population Names}
\usage{
ypr_population_names(populations)
}
\arguments{
\item{populations}{An object of class \code{\link[=ypr_populations]{ypr_populations()}}.}
}
\value{
A character vector of the unique population names.
}
\description{
Population Names
}
\examples{
ypr_population_names(ypr_populations(Rk = c(2.5, 3, 2.5), expand = FALSE))
}
\seealso{
Other populations: 
\code{\link{as_ypr_populations}()},
\code{\link{chilliwack_bt_05}},
\code{\link{ypr_plot_yield}()},
\code{\link{ypr_populations_expand}()},
\code{\link{ypr_populations_update}()},
\code{\link{ypr_populations}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}
}
\concept{populations}
