% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R, R/update.R
\name{ypr_population_update}
\alias{ypr_population_update}
\alias{update.ypr_population}
\title{Update Population Parameters}
\usage{
ypr_population_update(population, ...)

\method{update}{ypr_population}(object, ...)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}

\item{...}{One or more of the arguments from \code{ypr_population()}.}

\item{object}{The population or populations.}
}
\value{
An object of class \code{ypr_population}.
}
\description{
Updates an object of class \code{\link[=ypr_population]{ypr_population()}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{update}: Update Population Parameters
}}

\examples{
ypr_population_update(ypr_population(), Rk = 2.5)
}
