% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_populations}
\alias{ypr_populations}
\title{Populations}
\usage{
ypr_populations(..., expand = TRUE)
}
\arguments{
\item{...}{Unused parameters.}

\item{expand}{A flag specifying whether to expand parameter combinations.}
}
\value{
A list of \code{\link[=ypr_population]{ypr_population()}} objects
}
\description{
Populations
}
\examples{
ypr_populations(Rk = c(2.5, 4.6), Hm = c(0.2, 0.05))
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}}
}
