% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_populations_expand}
\alias{ypr_populations_expand}
\title{Expand Populations}
\usage{
ypr_populations_expand(populations)
}
\arguments{
\item{populations}{An object of class \code{\link[=ypr_populations]{ypr_populations()}}.}
}
\value{
An object of class \code{ypr_population}.
}
\description{
An object of class \code{\link[=ypr_population]{ypr_population()}} of all unique combinations of parameter values.
}
\examples{
ypr_populations_expand(ypr_populations(Rk = c(2.5, 4, 2.5), Hm = c(0.1, 0.2, 0.1)))
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}}
}
