% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R, R/update.R
\name{ypr_populations_update}
\alias{ypr_populations_update}
\alias{update.ypr_populations}
\title{Update Population Parameters}
\usage{
ypr_populations_update(populations, ...)

\method{update}{ypr_populations}(object, ...)
}
\arguments{
\item{populations}{An object of class \code{\link[=ypr_populations]{ypr_populations()}}.}

\item{...}{One or more of the arguments from \code{ypr_population()}.}

\item{object}{The population or populations.}
}
\value{
An object of class \code{ypr_population}.
}
\description{
Updates an object of class \code{\link[=ypr_population]{ypr_population()}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{update}: Update Populations Parameters
}}

\examples{
ypr_populations_update(ypr_populations(Rk = c(2.5, 4)), Rk = 2.5)
}
\seealso{
Other populations: 
\code{\link{as_ypr_populations}()},
\code{\link{chilliwack_bt_05}},
\code{\link{ypr_plot_yield}()},
\code{\link{ypr_populations_expand}()},
\code{\link{ypr_populations}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}
}
\concept{populations}
