% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{ypr_report}
\alias{ypr_report}
\title{Report}
\usage{
ypr_report(population, title = "Population Report", date = Sys.Date(),
  file = "report.Rmd")
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}

\item{title}{A string of the report title.}

\item{date}{A date of the report date.}

\item{file}{A string of the file name.}
}
\value{
An invisible character vector of the contents of the file.
}
\description{
Creates an Rmd file that can be used to generate a report.
}
\examples{
\dontrun{
cat(ypr_report(ypr_population(), file = tempfile()), sep = "\\n")
}
}
