% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{ypr_sr}
\alias{ypr_sr}
\title{Stock-Recruitment Parameters}
\usage{
ypr_sr(population)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}
}
\value{
A named numeric vector of alpha, beta and R0F (the number of recruits at the fished equilibrium )
}
\description{
Stock-Recruitment Parameters
}
\examples{
ypr_sr(ypr_population()) # Beverton-Holt
ypr_sr(ypr_population(BH = 0L)) # Ricker
}
