% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship.R
\name{ypr_survivorship}
\alias{ypr_survivorship}
\title{Calculate Survivorship}
\usage{
ypr_survivorship(t, nu = 0.2, Rt = 1L)
}
\arguments{
\item{t}{The age.}

\item{nu}{The annual interval natural mortality rate.}

\item{Rt}{A count of the recruitment age for the stock recruitment relationship.}
}
\value{
The expected survivorship.
}
\description{
Calculates the expected survivorship to age t.
}
\examples{
plot(0:10, ypr_survivorship(0:10), type = "l")
}
