% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_biomass}
\alias{ypr_tabulate_biomass}
\title{Table Fish}
\usage{
ypr_tabulate_biomass(population)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}
}
\value{
A data frame
}
\description{
Produces a data frame of the 'Weight' and 'Fecundity' and the number of
'Survivors' and 'Spawners' and the total 'Biomass' and 'Eggs' by
'Age' class.
}
\examples{
ypr_tabulate_biomass(ypr_population())
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}} and \code{\link[=ypr_plot_fish]{ypr_plot_fish()}}
}
