% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate-parameters.R
\name{ypr_tabulate_parameters}
\alias{ypr_tabulate_parameters}
\title{Tabulate Population Parameters}
\usage{
ypr_tabulate_parameters(population)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}
}
\value{
A table of population parameters
}
\description{
Tabulate Population Parameters
}
\examples{
ypr_tabulate_parameters(ypr_population())
}
\seealso{
Other tabulate: 
\code{\link{ypr_detabulate_parameters}()},
\code{\link{ypr_report}()},
\code{\link{ypr_tabulate_biomass}()},
\code{\link{ypr_tabulate_fish}()},
\code{\link{ypr_tabulate_schedule}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}

Other parameters: 
\code{\link{ypr_detabulate_parameters}()}
}
\concept{parameters}
\concept{tabulate}
