% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_sr}
\alias{ypr_tabulate_sr}
\title{Stock-Recruitment Parameters}
\usage{
ypr_tabulate_sr(population, Ly = 0, harvest = FALSE, biomass = FALSE)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for just harvested fish or the total number of captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}
}
\value{
A table of stock-recruitment parameters.
}
\description{
Stock-Recruitment Parameters
}
\examples{
ypr_tabulate_sr(ypr_population()) # Beverton-Holt
ypr_tabulate_sr(ypr_population(BH = 0L)) # Ricker
}
