% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_sr.ypr_populations}
\alias{ypr_tabulate_sr.ypr_populations}
\title{Stock-Recruitment Parameters}
\usage{
\method{ypr_tabulate_sr}{ypr_populations}(object, Ly = 0,
  harvest = TRUE, biomass = FALSE, all = FALSE, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for harvested fish or captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}

\item{all}{A flag indicating whether to include all parameter values.}

\item{...}{Unused parameters.}
}
\value{
A table of stock-recruitment parameters.
}
\description{
Stock-Recruitment Parameters
}
\examples{
ypr_tabulate_sr(ypr_populations(Rk = c(2.5, 4.6)))
}
