% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_yield.ypr_population}
\alias{ypr_tabulate_yield.ypr_population}
\title{Tabulate Yield}
\usage{
\method{ypr_tabulate_yield}{ypr_population}(object, Ly = 0,
  harvest = FALSE, biomass = FALSE, optimal = TRUE, all = FALSE,
  ...)
}
\arguments{
\item{object}{The populations or populations to tabulate.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for just harvested fish or the total number of captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}

\item{optimal}{A flag indicating whether to include the optimal yield.}

\item{all}{A flag indicating whether to include all parameter values.}

\item{...}{Unused parameters.}
}
\value{
A data frame.
}
\description{
Tabulate Yield
}
\examples{
ypr_tabulate_yield(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_yield}}
}
