% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_yields.ypr_populations}
\alias{ypr_tabulate_yields.ypr_populations}
\title{Tabulate Yields}
\usage{
\method{ypr_tabulate_yields}{ypr_populations}(object, pi = seq(0, 1,
  length.out = 100), Ly = 0, harvest = TRUE, biomass = FALSE,
  all = FALSE, ..., chk = TRUE)
}
\arguments{
\item{object}{The populations or populations.}

\item{pi}{A vector of probabilities of capture to calculate the yield for.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for harvested fish or captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}

\item{all}{A flag indicating whether to include all parameter values.}

\item{...}{Unused parameters.}

\item{chk}{A flag specifying whether to check the other parameters.}
}
\value{
A data frame.
}
\description{
Tabulate Yields
}
\examples{
ypr_tabulate_yields(ypr_populations(Rk = c(3, 5)), pi = seq(0, 1, length.out = 10))
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}} and \code{\link[=ypr_yield]{ypr_yield()}}
}
