% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability.R
\name{ypr_vulnerability}
\alias{ypr_vulnerability}
\title{Calculate Vulnerability}
\usage{
ypr_vulnerability(L, Llo = 50, Lup = 100)
}
\arguments{
\item{L}{The length.}

\item{Llo}{The lower harvest length.}

\item{Lup}{The upper harvest length.}
}
\value{
The expected vulnerability to fishing
}
\description{
Calculates the vulnerability to fishing for an individual of length L with
the length slot Llo to Lup and no hooking mortality.
}
\examples{
plot(0:100, ypr_vulnerability(0:100), type = "l")
}
