% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{ypr_weight}
\alias{ypr_weight}
\title{Calculate Weight}
\usage{
ypr_weight(L, a = 1e-05, b = 3)
}
\arguments{
\item{L}{The length.}

\item{a}{The condition factor.}

\item{b}{The scaling exponent.}
}
\value{
The expected weight for an individual of length L.
}
\description{
Calculates the expected weight for an individual of length L based on
the allometric equation.
}
\examples{
plot(0:100, ypr_weight(0:100), type = "l")
}
