% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{ypr_yield}
\alias{ypr_yield}
\title{Yield}
\usage{
ypr_yield(population, Ly = 0, harvest = TRUE, biomass = TRUE,
  sanitize = TRUE, check = TRUE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{Ly}{The minimum length fish to consider.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{sanitize}{A flag indicating whether to replace negative or NaNs with 0.}

\item{check}{A flag indicating whether to check the arguments.}
}
\description{
Yield
}
\examples{
ypr_yield(ypr_population())
}
