% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield-limits.R
\name{ypr_yield_limits}
\alias{ypr_yield_limits}
\title{Yield with Limits}
\usage{
ypr_yield_limits(x, x2, Ly = 0, harvest = TRUE, biomass = TRUE,
  sanitize = TRUE, check = TRUE, n = 10^4, level = 0.95)
}
\arguments{
\item{x}{A list of population life-history parameters.}

\item{x2}{A second list of population life-history parameters.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{sanitize}{A flag indicating whether to replace negative or NaNs with 0.}

\item{check}{A flag indicating whether to check the arguments.}

\item{n}{A numeric of the number of iterations.}

\item{level}{A probability of the interval.}
}
\value{
A double vector of the yield and its lower and upper limits.
}
\description{
Calculates the yield with lower and upper limits for a population.
The uncertainty in each parameter is assumed to be a uniform distribution between the value in x and the value in x2.
The parameters are assumed to be uncorrelated and there is no requirement
for the value in x to be less than the value in x2
The parameters \code{tmax} and \code{Rk} must not vary.
}
\examples{
population <- ypr_population()
population$Rk <- 10
ypr_yield_limits(population, ypr_population(), 2)
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_yield}}
}
