% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yields.R
\name{ypr_yields}
\alias{ypr_yields}
\title{Yields}
\usage{
ypr_yields(population, mu = seq(0, 1, length.out = 100), Ly = 0,
  harvest = TRUE, biomass = TRUE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{mu}{A vector of probabilities of capture to calculate the yield for.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}
}
\description{
Calculates the yield(s) for a population based on one or more capture rates.
}
\examples{
mu <- seq(0, 1, length.out = 100)
plot(mu, ypr_yields(ypr_population(), mu), type = "l")
}
\seealso{
\code{\link{ypr_population}}, \code{\link{ypr_optimize}},
\code{\link{ypr_yield}} and \code{\link{ypr_plot}}
}
