% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yields-uncertainty.R
\name{ypr_yields_uncertainty}
\alias{ypr_yields_uncertainty}
\title{Yields}
\usage{
ypr_yields_uncertainty(population, population2, mu = seq(0, 1, length.out =
  100), Ly = 0, harvest = TRUE, biomass = TRUE, n = 10^4,
  level = 0.95, mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{population2}{A second list of population life-history parameters.}

\item{mu}{A vector of probabilities of capture to calculate the yield for.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{n}{A numeric of the number of iterations.}

\item{level}{A probability of the interval.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously.
The option is initialized from environment variable MC_CORES if set.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
A double vector of the yields.
}
\description{
Calculates the yield(s) for a population based on one or more capture rates.
}
\examples{
population <- ypr_population()
population$Rk <- 10
ypr_yields_uncertainty(ypr_population(), population, n = 10)
}
\seealso{
\code{\link{ypr_population}} and
\code{\link{ypr_yields}}
}
