% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-ecotype.R
\name{as_ypr_ecotypes}
\alias{as_ypr_ecotypes}
\title{Ecotype Parameters}
\usage{
as_ypr_ecotypes(populations, weights)
}
\arguments{
\item{populations}{A list with all parameters for each ecotype in the
population}

\item{weights}{A numeric vector giving the weighting for how much it
contributes to the overall population}
}
\description{
Creates an object of class ecotype.
}
