% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-ecotypes.R
\name{as_ypr_ecotypes}
\alias{as_ypr_ecotypes}
\title{Coerce to an Ecotypes Object}
\usage{
as_ypr_ecotypes(
  populations = ypr_populations(Linf = c(100, 1000)),
  weights = c(0.5, 0.5),
  names = NULL
)
}
\arguments{
\item{populations}{A list with each element providing the population
parameters for each ecotype.}

\item{weights}{A numeric vector giving the relative proportion of recruits belonging
to each ecotype.}

\item{names}{An optional character vector providing names for each ecotype.}
}
\description{
Creates an object of class ypr_ecotypes from a list of ypr_populations.
}
\examples{
as_ypr_ecotypes()
as_ypr_ecotypes(ypr_populations(Ls = c(10, 15, 20)), c(1, 2, 1), c("river", "deep", "shallow"))
}
\seealso{
Other ecotypes: 
\code{\link{ypr_ecotypes_update}()},
\code{\link{ypr_ecotypes}()}
}
\concept{ecotypes}
