% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-ecotypes.R
\name{as_ypr_ecotypes}
\alias{as_ypr_ecotypes}
\alias{as_ypr_ecotypes.ypr_population}
\alias{as_ypr_ecotypes.ypr_populations}
\title{Coerce to an Ecotypes Object}
\usage{
as_ypr_ecotypes(x, ...)

\method{as_ypr_ecotypes}{ypr_population}(x, ...)

\method{as_ypr_ecotypes}{ypr_populations}(x, names = NULL, ...)
}
\arguments{
\item{x}{The object to coerce.}

\item{...}{Additional arguments.}

\item{names}{A character vector providing names for each ecotype.}
}
\value{
An object of class ypr_ecotypes.
}
\description{
Coerce to an Ecotypes Object
}
\section{Methods (by class)}{
\itemize{
\item \code{ypr_population}: Coerce a Population Object to an Ecotypes Object

\item \code{ypr_populations}: Coerce a Populations Object to an Ecotypes Object
}}

\examples{
as_ypr_ecotypes(ypr_population())
as_ypr_ecotypes(ypr_populations(Ls = c(10, 15, 20)), c("river", "deep", "shallow"))
}
\seealso{
Other ecotypes: 
\code{\link{ypr_ecotypes_update}()},
\code{\link{ypr_ecotypes}()}
}
\concept{ecotypes}
