% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-ecotypes.R
\name{as_ypr_ecotypes}
\alias{as_ypr_ecotypes}
\title{Ecotype Parameters}
\usage{
as_ypr_ecotypes(
  populations = ypr_populations(Linf = c(100, 1000)),
  weights = c(0.5, 0.5),
  names = NULL
)
}
\arguments{
\item{populations}{A list with all parameters for each ecotype in the
population}

\item{weights}{A numeric vector giving the weighting for how much it
contributes to the overall population}

\item{names}{A character vector providing names for each ecotype. Default is
NULL.}
}
\description{
Creates an object of class ecotype.
}
\examples{
as_ypr_ecotypes()
as_ypr_ecotypes(ypr_populations(Ls = c(10, 15, 20)), c(1, 2, 1), c("river", "deep", "shallow"))
}
\seealso{
Other ecotypes: 
\code{\link{update.ypr_ecotypes}()},
\code{\link{ypr_ecotypes}()}
}
\concept{ecotypes}
