% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-population.R
\name{as_ypr_population}
\alias{as_ypr_population}
\title{Coerce to a Population Object}
\usage{
as_ypr_population(x)
}
\arguments{
\item{x}{The object to coerce.}
}
\description{
Creates an object of class ypr_population from a data frame of population parameters.
}
\examples{
as_ypr_population(as.data.frame(ypr_population()))
}
