% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ypr-populations.R
\name{as_ypr_populations}
\alias{as_ypr_populations}
\title{Coerce to a Populations Object}
\usage{
as_ypr_populations(x)
}
\arguments{
\item{x}{The object to coerce.}
}
\description{
Creates an object of class ypr_population from a data frame of populations parameters.
}
\examples{
as_ypr_populations(as.data.frame(ypr_populations(Rk = c(3, 4))))
}
\seealso{
Other populations: 
\code{\link{chilliwack_bt_05}},
\code{\link{update.ypr_populations}()},
\code{\link{ypr_plot_yield}()},
\code{\link{ypr_population_names}()},
\code{\link{ypr_populations_expand}()},
\code{\link{ypr_populations}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}
}
\concept{populations}
