% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot.ypr_population}
\alias{plot.ypr_population}
\title{Plot Population Schedule}
\usage{
\method{plot}{ypr_population}(x, type = "b", ...)
}
\arguments{
\item{x}{The population to plot.}

\item{type}{1-character string giving the type of plot desired.  The
    following values are possible, for details, see \code{\link[graphics]{plot}}:
    \code{"p"} for points, \code{"l"} for lines,
    \code{"b"} for both points and lines,
    \code{"c"} for empty points joined by lines,
    \code{"o"} for overplotted points and lines,
    \code{"s"} and \code{"S"} for stair steps and
    \code{"h"} for histogram-like vertical lines.  Finally,
    \code{"n"} does not produce any points or lines.}

\item{...}{Additional arguments passed to graphics::plot.formula functions.}
}
\value{
An invisible copy of the original object.
}
\description{
Plot Population Schedule
}
\examples{
\dontrun{
plot(ypr_population())
}
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}} and \code{\link[=ypr_tabulate_schedule]{ypr_tabulate_schedule()}}
}
