% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quesnel_lt}
\alias{quesnel_lt}
\title{Quesnel Lake Lake Trout Population Parameters}
\format{An object of class \code{\link{ypr_population}}.}
\usage{
quesnel_lt
}
\description{
The population parameters for
Lake Trout in Quesnel Lake, BC.
}
\details{
The estimates are liable to change and should not be used for management.
}
\examples{
quesnel_lt
ypr_plot_yield(quesnel_lt)
}
\keyword{datasets}
