% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr-package.R
\docType{package}
\name{ypr-package}
\alias{ypr}
\alias{ypr-package}
\title{ypr: Yield Per Recruit for Recreational Fisheries}
\description{
An implementation of yield per recruit methods for
recreational fisheries in R.
Given a list of life history parameters it can generate a data frame
of the schedule and find the capture rate which maximises the yield.
The yield, which is provided in terms of the proportion of the unfished population,
can be based on the total biomass vs number of fish;
harvested vs captured and large (trophy) versus all fish.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/ypr}
  \item Report bugs at \url{https://github.com/poissonconsulting/ypr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
