% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr-package.R
\docType{package}
\name{ypr-package}
\alias{ypr}
\alias{ypr-package}
\title{ypr: Yield Per Recruit}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

An implementation of equilibrium-based yield per recruit methods.
    Yield per recruit methods can used to estimate the optimal yield for a fish population.
    The yield can be based on the number of fish caught (or harvested) or 
    biomass caught for all fish or just large (trophy) individuals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/ypr}
  \item Report bugs at \url{https://github.com/poissonconsulting/ypr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
