\name{YPR-package}
\alias{YPR-package}
\alias{YPR}
\docType{package}
\title{
Yield Per Recruit (YPR) model.
}
\description{
For now, this is a length based Yield Per Recruit (YPR) model.
}
\details{
\tabular{ll}{
Package: \tab YPR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-10-03\cr
License: \tab LGPL >= 3.0\cr
LazyLoad: \tab yes\cr
}

}
\author{
Benoit Bruneau

Maintainer: Benoit Bruneau <benoit.bruneau1@gmail.com>
}

