\name{YPR}
\alias{YPR}
\docType{package}
\title{
Yield Per Recruit (YPR) model.
}
\description{
This is a length-based Yield Per Recruit (YPR) model. It will soon include a weight-based YPR component.
}
\details{
\tabular{ll}{
Package: \tab YPR\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2011-11-18\cr
License: \tab LGPL >= 3.0\cr
LazyLoad: \tab yes\cr
}

}
\author{
Benoit Bruneau,\cr
François Grégoire and\cr
Diane Archambault

Maintainer: Benoit Bruneau <benoit.bruneau1@gmail.com>
}

