% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr.R
\name{ypr}
\alias{ypr}
\title{Yield-Per-Recruit}
\usage{
ypr(tmax = 20L, k = 0.1, Linf = 100, t0 = 0, a = 1e-05, b = 3,
  Llo = Linf/2, Lup = Linf, mu = 0.2, Lm = Linf/2, fa = a, fb = b)
}
\arguments{
\item{tmax}{The maximum age.}

\item{k}{The growth coefficient.}

\item{Linf}{The mean maximum length.}

\item{t0}{The age at length 0.}

\item{a}{The condition factor.}

\item{b}{The scaling exponent.}

\item{Llo}{The lower harvest length.}

\item{Lup}{The upper harvest length.}

\item{mu}{The annual interval mortality rate.}

\item{Lm}{The length at maturity.}

\item{fa}{The fecundity factor.}

\item{fb}{The fecundity scaling exponent.}
}
\description{
Yield-Per-Recruit
}
\examples{
ypr()
}
