% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr.R
\name{ypr}
\alias{ypr}
\title{Yield-Per-Recruit}
\usage{
ypr(k, Linf = 1000, t0 = 0, alpha = 1, beta = 3, Lm = 500, Lv = 400)
}
\arguments{
\item{k}{A number of the growth coefficient.}

\item{Linf}{A number of the mean maximum length.}

\item{t0}{A number}

\item{alpha}{A number of the}

\item{beta}{A number of the}

\item{Lm}{A number of the length at maturity.}

\item{Lv}{A number of the length at vulnerable to capture.}
}
\description{
The default values are for large piscivorous Rainbow Trout.
}
