% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypr_age_at_length}
\alias{ypr_age_at_length}
\title{Age At Length}
\usage{
ypr_age_at_length(population, length)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}

\item{length}{A numeric vector of the length (cm).}
}
\value{
A double vector of the lengths.
}
\description{
Age At Length
}
\examples{
ypr_age_at_length(ypr_population(), seq(0, 100, by = 10))
}
\seealso{
Other calculate: 
\code{\link{ypr_exploitation}()},
\code{\link{ypr_length_at_age}()},
\code{\link{ypr_optimize}()},
\code{\link{ypr_yields}()},
\code{\link{ypr_yield}()}
}
\concept{calculate}
