% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detabulate.R
\name{ypr_detabulate_parameters}
\alias{ypr_detabulate_parameters}
\title{Detabulate Population Parameters}
\usage{
ypr_detabulate_parameters(x)
}
\arguments{
\item{x}{A data frame with columns Parameter and Value specifying one or more
parameters and their values.}
}
\value{
An object of class \code{\link[=ypr_population]{ypr_population()}}
}
\description{
Detabulate Population Parameters
}
\examples{
ypr_detabulate_parameters(ypr_tabulate_parameters(ypr_population()))
}
\seealso{
Other tabulate: 
\code{\link{ypr_report}()},
\code{\link{ypr_tabulate_biomass}()},
\code{\link{ypr_tabulate_fish}()},
\code{\link{ypr_tabulate_parameters}()},
\code{\link{ypr_tabulate_schedule}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}

Other parameters: 
\code{\link{ypr_tabulate_parameters}()}
}
\concept{parameters}
\concept{tabulate}
