% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_ecotypes}
\alias{ypr_ecotypes}
\title{Ecotypes}
\usage{
ypr_ecotypes(...)
}
\arguments{
\item{...}{One or more of the arguments from \code{ypr_population()}.}
}
\value{
A list of \code{\link{ypr_population}} objects each representing
a unique ecotype
with identical recruitment (\code{tR} \code{BH} and \code{Rk}) and
fishery (\code{Llo}, \code{Lup}, \code{NC}, \code{pi} and \code{q}) parameters.
The relative proportion of the population belonging to each ecotype is indicated
by the \code{Rmax} parameters.

An object of class \code{ypr_ecotypes}.
}
\description{
Ecotypes
}
\examples{
ypr_populations(Rk = c(2.5, 4.6), Hm = c(0.2, 0.05))
}
