% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_ecotypes}
\alias{ypr_ecotypes}
\title{Ecotypes}
\usage{
ypr_ecotypes(..., weights)
}
\arguments{
\item{...}{Unused parameters.}

\item{weights}{A numeric vector giving the weighting for how much it
contributes to the overall population}
}
\value{
A list of \code{\link[=ypr_ecotypes]{ypr_ecotypes()}} objects
}
\description{
Ecotypes
}
\examples{
ypr_ecotypes()
}
\seealso{
Other ecotypes: 
\code{\link{as_ypr_ecotypes}()}
}
\concept{ecotypes}
