% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotypes.R
\name{ypr_ecotypes}
\alias{ypr_ecotypes}
\title{Create Ecotypes Object}
\usage{
ypr_ecotypes(..., weights = NULL, names = NULL)
}
\arguments{
\item{...}{Unused parameters.}

\item{weights}{A numeric vector giving the weighting for how much it
contributes to the overall population.}

\item{names}{A character vector of unique ecotype names.}
}
\value{
A list of \code{\link[=ypr_ecotypes]{ypr_ecotypes()}} objects
}
\description{
Creates an object of class ecotypes.
}
\examples{
ypr_ecotypes(Linf = c(1, 2), weights = c(1, 1))
ypr_ecotypes(Linf = c(1, 2), t0 = c(0, 0.5), weights = c(1, 1))
}
\seealso{
Other ecotypes: 
\code{\link{as_ypr_ecotypes}()},
\code{\link{update.ypr_ecotypes}()}
}
\concept{ecotypes}
