% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R, R/update.R
\name{ypr_ecotypes_update}
\alias{ypr_ecotypes_update}
\alias{update.ypr_ecotypes}
\title{Update Ecotypes Parameters}
\usage{
ypr_ecotypes_update(ecotypes, ...)

\method{update}{ypr_ecotypes}(object, ...)
}
\arguments{
\item{ecotypes}{An object of class \code{\link[=ypr_ecotypes]{ypr_ecotypes()}}.}

\item{...}{One or more of the arguments from \code{ypr_population()}.}

\item{object}{The population or populations.}
}
\value{
An object of class \code{ypr_ecotypes}.
}
\description{
Allows updates to a parameter of an object of class \code{\link[=ypr_ecotypes]{ypr_ecotypes()}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{update}: Update Ecotypes Parameters
}}

\examples{
ypr_ecotypes_update(ypr_ecotypes(Linf = c(2.5, 4), weights = c(1, 1)), Linf = 2.5)
}
\seealso{
Other ecotypes: 
\code{\link{as_ypr_ecotypes}()},
\code{\link{ypr_ecotypes}()}
}
\concept{ecotypes}
