% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{ypr_names}
\alias{ypr_names}
\alias{ypr_names.ypr_population}
\alias{ypr_names.ypr_populations}
\alias{ypr_names.ypr_ecotypes}
\title{Population(s) or Ecotype Names}
\usage{
ypr_names(x, ...)

\method{ypr_names}{ypr_population}(x, ...)

\method{ypr_names}{ypr_populations}(x, ...)

\method{ypr_names}{ypr_ecotypes}(x, ...)
}
\arguments{
\item{x}{An object of class ypr_population, ypr_populations or ypr_ecotypes.}

\item{...}{Unused.}
}
\value{
A character vector of the unique parameter based names.
}
\description{
Generates set of unique names based on differences in parameter values.
}
\details{
Parameter RPR is ignored because it is irrelevant to population(s) and does not
distinguish between ecotypes.
}
\section{Methods (by class)}{
\itemize{
\item \code{ypr_names(ypr_population)}: Population Names

\item \code{ypr_names(ypr_populations)}: Populations Names

\item \code{ypr_names(ypr_ecotypes)}: Ecotypes Names

}}
\examples{
ypr_names(ypr_population())
ypr_names(ypr_populations())
ypr_names(ypr_populations())
}
