% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{ypr_optimize}
\alias{ypr_optimize}
\alias{ypr_optimise}
\title{Optimize Capture}
\usage{
ypr_optimize(population, Ly = 0, harvest = FALSE, biomass = FALSE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}
}
\description{
Finds the capture rate that maximises the yield for a given population.
}
\examples{
ypr_optimize(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_yields}}
}
