% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{ypr_optimize}
\alias{ypr_optimize}
\alias{ypr_optimise}
\title{Optimize Capture}
\usage{
ypr_optimize(population, Ly = 0, harvest = TRUE, biomass = FALSE)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for harvested fish or captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}
}
\value{
The interval annual capture probability (pi) that maximises the yield.
}
\description{
Finds the interval annual capture probability (pi) that maximises the yield for a given population.
}
\examples{
ypr_optimize(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_yield}}
}
