% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ypr_plot}
\alias{ypr_plot}
\title{Plot Population Yields by Capture Probabilities}
\usage{
ypr_plot(population, pi = seq(0, 1, length.out = 100), Ly = 0,
  harvest = FALSE, biomass = FALSE)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}

\item{pi}{A vector of probabilities of capture to calculate the yield for.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for just harvested fish or the total number of captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}
}
\description{
Plots the population yield by the capture probabilities (pi).
The actual and optimal capture probabilities are indicated by
blue dashed and red dotted lines, respectively.
}
\examples{
\dontrun{
ypr_plot(ypr_population())
}
}
\seealso{
\code{\link{ypr_population}}, \code{\link{ypr_yields}}
and \code{\link{ypr_optimize}}
}
