% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ypr_plot}
\alias{ypr_plot}
\title{Plot Yields by Capture Probabilities}
\usage{
ypr_plot(population, mu = seq(0, 0.5, length.out = 30), Ly = 0,
  harvest = TRUE, biomass = TRUE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{mu}{A vector of probabilities of capture to calculate the yield for.}

\item{Ly}{The minimum length fish to consider.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}
}
\description{
Plot Yields by Capture Probabilities
}
\examples{
\dontrun{
ypr_plot(ypr_population())
}
}
