% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-schedule.R
\name{ypr_plot_schedule}
\alias{ypr_plot_schedule}
\title{Plot Population or Ecotypes Schedule Terms}
\usage{
ypr_plot_schedule(population, x = "Age", y = "Length")
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}

\item{x}{A string of the term on the x-axis.}

\item{y}{A string of the term on the y-axis.}
}
\value{
A ggplot2 object.
}
\description{
Produces a bivariate line plot of two schedule terms.
}
\examples{
ypr_plot_schedule(ypr_population())
}
\seealso{
Other schedule: 
\code{\link{ypr_tabulate_schedule}()}

Other plot: 
\code{\link{ypr_plot_biomass}()},
\code{\link{ypr_plot_fish}()},
\code{\link{ypr_plot_sr}()},
\code{\link{ypr_plot_yield}()}
}
\concept{plot}
\concept{schedule}
