% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ypr_plot_schedule}
\alias{ypr_plot_schedule}
\title{Plot Population Schedule Terms}
\usage{
ypr_plot_schedule(population, x = "Age", y = "Length")
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}

\item{x}{A string of the term on the x-axis.}

\item{y}{A string of the term on the y-axis.}
}
\value{
A ggplot2 object.
}
\description{
Produces a bivariate line plot of two schedule terms.
}
\examples{
ypr_plot_schedule(ypr_population())
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}} and \code{\link[=ypr_tabulate_schedule]{ypr_tabulate_schedule()}}
}
