% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ypr_plot_yield}
\alias{ypr_plot_yield}
\title{Plot Yield by Capture}
\usage{
ypr_plot_yield(object, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{...}{Unused parameters.}
}
\value{
A ggplot2 object.
}
\description{
Plots the 'Yield', 'Age', 'Length', 'Weight', 'Effort', or 'YPUE'
by the annual interval capture probability.
}
\examples{
ypr_plot_yield(ypr_population())
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}} and \code{\link[=ypr_yields]{ypr_yields()}}
}
