% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_population}
\alias{ypr_population}
\title{Population Parameters}
\usage{
ypr_population(tmax = 20L, k = 0.15, Linf = 100, t0 = 0, b = 3,
  Lm = Linf/2, fb = 1, Rt = 1L, nu = 0.2, Lv = Linf/2, Vp = 5,
  Llo = 0, Lup = Linf, Nc = 0, mu = 0.2, rho = 0, eta = 0, Rk = 5)
}
\arguments{
\item{tmax}{The maximum age.}

\item{k}{The growth coefficient.}

\item{Linf}{The mean maximum length.}

\item{t0}{The extrapolated age at length 0.}

\item{b}{The condition scaling exponent.}

\item{Lm}{The length at maturity.}

\item{fb}{The fecundity scaling exponent.}

\item{Rt}{The recruitment age for the stock recruitment relationship.}

\item{nu}{The conditional annual probability of dying due to natural causes.}

\item{Lv}{The length at which 50\% vulnerable to harvest.}

\item{Vp}{The vulnerability to harvest power.}

\item{Llo}{The lower harvest slot length.}

\item{Lup}{The upper harvest slot length.}

\item{Nc}{The slot limits non-compliance probability.}

\item{mu}{The conditional annual probability of being captured.}

\item{rho}{The release probability.}

\item{eta}{The handling mortality probability.}

\item{Rk}{The numbers of spawners per spawner at low density.}
}
\value{
A list of the life-history parameters for a particular population.
}
\description{
Generates a list of the life-history parameters
for a fish population.
The default parameters are for a hypothetical population that
lives to 20 years of age,
grows to a mean maximum length of 100 with a k of 0.15,
undergoes allometric growth and fecundity and matures at a length of 50
has a natural mortality rate of 20% from age 1
and has a capture rate of 20% from a length of 50 onwards
with no catch and release.
}
\examples{
ypr_population()
}
\seealso{
\code{\link{ypr_schedule}}, \code{\link{plot.ypr_population}},
\code{\link{ypr_yield}}, \code{\link{ypr_plot}}
and \code{\link{ypr_optimize}}
}
