% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_population}
\alias{ypr_population}
\title{Population Parameters}
\usage{
ypr_population(tmax = 20L, k = 0.15, Linf = 100, t0 = 0, b = 3,
  Lm = Linf/2, fb = 1, Rt = 1L, nu = 1 - exp(-1.5 * k), Lv = Linf/2,
  Vp = 100, Llo = 0, Lup = Linf, Nc = 0, mu = 0.2, rho = 0,
  eta = 0, Rk = 5, R0 = 1)
}
\arguments{
\item{tmax}{The maximum age.}

\item{k}{The growth coefficient.}

\item{Linf}{The mean maximum length.}

\item{t0}{The extrapolated age at length 0.}

\item{b}{The condition scaling exponent.}

\item{Lm}{The length at maturity.}

\item{fb}{The fecundity scaling exponent.}

\item{Rt}{The recruitment age for the stock recruitment relationship.}

\item{nu}{The conditional annual probability of dying due to natural causes.}

\item{Lv}{The length at which 50\% vulnerable to harvest.}

\item{Vp}{The vulnerability to harvest power.}

\item{Llo}{The lower harvest slot length.}

\item{Lup}{The upper harvest slot length.}

\item{Nc}{The slot limits non-compliance probability.}

\item{mu}{The conditional annual probability of being captured.}

\item{rho}{The release probability.}

\item{eta}{The hooking mortality probability.}

\item{Rk}{The numbers of spawners per spawner at low density.}

\item{R0}{The number of recruits at the carrying capacity.}
}
\value{
An object of class \code{ypr_population}.
}
\description{
Generates a list of the life-history parameters
for a fish population.
}
\details{
The default parameters are for a hypothetical population that
allometrically grows to a mean maximum length of 100 with a k of 0.15,
and both matures and becomes vulnerable to harvest at a length of 50.
}
\examples{
ypr_population()
}
\seealso{
\code{\link{plot.ypr_population}} and \code{\link{ypr_population_update}}.
}
