% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names-deprec.R
\name{ypr_population_names}
\alias{ypr_population_names}
\title{Population Names}
\usage{
ypr_population_names(population)
}
\arguments{
\item{population}{An object of class ypr_population, ypr_populations or ypr_ecotypes.}
}
\value{
A character vector of the unique parameter based names.
}
\description{
Generates set of unique names based on differences in parameter values.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\seealso{
\code{\link[=ypr_names]{ypr_names()}}
}
