% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_population_names}
\alias{ypr_population_names}
\title{Population Names}
\usage{
ypr_population_names(populations)
}
\arguments{
\item{populations}{An object of class \code{\link[=ypr_populations]{ypr_populations()}}.}
}
\value{
A character vector of the unique population names.
}
\description{
Population Names
}
\examples{
ypr_population_names(ypr_populations(Rk = c(2.5, 3, 2.5), expand = FALSE))
}
