% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-deprec.R
\name{ypr_population_update}
\alias{ypr_population_update}
\title{Update a Population Object}
\usage{
ypr_population_update(population, ...)
}
\arguments{
\item{population}{A ypr_population object.}

\item{...}{One or more parameter values from \code{ypr_population()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} for \code{\link[=ypr_update]{ypr_update()}}.
}
