% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_population_update}
\alias{ypr_population_update}
\title{Update Population Parameters}
\usage{
ypr_population_update(population, ...)
}
\arguments{
\item{population}{An object of class \code{ypr_population}}

\item{...}{One or more of the arguments from \code{ypr_population()}.}
}
\value{
An object of class \code{ypr_population}.
}
\description{
Updates an object of class \code{\link{ypr_population}}.
}
\examples{
ypr_population_update(ypr_population(), Rk = 2.5)
}
\seealso{
\code{\link{ypr_population}}
}
