% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ypr_populations}
\alias{ypr_populations}
\title{Populations}
\usage{
ypr_populations(...)
}
\arguments{
\item{...}{One or more of the arguments from \code{ypr_population()}.}
}
\value{
A list of \code{\link{ypr_population}} objects
}
\description{
Populations
}
\examples{
ypr_populations(Rk = c(2.5, 4.6), Hm = c(0.2, 0.05))
}
\seealso{
\code{\link{ypr_population}}
}
