% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{ypr_populations}
\alias{ypr_populations}
\title{Populations}
\usage{
ypr_populations(..., expand = TRUE, names = NULL)
}
\arguments{
\item{...}{Unused parameters.}

\item{expand}{A flag specifying whether to expand parameter combinations.}

\item{names}{A character vector of unique ecotype names.}
}
\value{
A list of \code{\link[=ypr_population]{ypr_population()}} objects
}
\description{
Populations
}
\examples{
ypr_populations(Rk = c(2.5, 4.6), Hm = c(0.2, 0.05))
}
\seealso{
Other populations: 
\code{\link{as_ypr_populations}()},
\code{\link{chilliwack_bt_05}},
\code{\link{ypr_plot_yield}()},
\code{\link{ypr_populations_expand}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}
}
\concept{populations}
