% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/population.R
\name{update.ypr_populations}
\alias{update.ypr_populations}
\alias{ypr_populations_update}
\title{Update Population Parameters}
\usage{
\method{update}{ypr_populations}(object, ...)

ypr_populations_update(populations, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{...}{One or more of the arguments from \code{ypr_population()}.}

\item{populations}{An object of class \code{\link[=ypr_populations]{ypr_populations()}}.}
}
\value{
An object of class \code{ypr_population}.
}
\description{
Updates an object of class \code{\link[=ypr_population]{ypr_population()}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{update}: Update Populations Parameters
}}

\examples{
ypr_populations_update(ypr_populations(Rk = c(2.5, 4)), Rk = 2.5)
}
\seealso{
\code{\link[=ypr_population]{ypr_population()}}
}
