% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{ypr_schedule}
\alias{ypr_schedule}
\alias{ypr_schedule.ypr_population}
\alias{ypr_schedule.ypr_ecotypes}
\title{Life-History Schedule}
\usage{
ypr_schedule(population, ...)

\method{ypr_schedule}{ypr_population}(population, ...)

\method{ypr_schedule}{ypr_ecotypes}(population, ...)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}} or \code{\link{ypr_ecotypes}}.}

\item{...}{Unused}
}
\value{
A tibble of the life-history schedule by age and possibly ecotype.
}
\description{
Generates the life-history schedule by age for a population.
}
\section{Methods (by class)}{
\itemize{
\item \code{ypr_population}: Life-History Schedule Population

\item \code{ypr_ecotypes}: Life-History Schedule Ecotypes
}}

\examples{
ypr_schedule(ypr_population())
}
\seealso{
\code{\link{ypr_population}}
}
