% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{ypr_schedule}
\alias{ypr_schedule}
\title{Life-History Schedule}
\usage{
ypr_schedule(population, complete = FALSE, check = TRUE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{complete}{A flag indicating whether to generate a complete schedule.}

\item{check}{A flag indicating whether to check the arguments.}
}
\value{
A tibble of the life-history schedule by age.
}
\description{
Life-History Schedule
}
\examples{
ypr_schedule(ypr_population())
}
