% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{ypr_schedule}
\alias{ypr_schedule}
\title{Life-History Schedule}
\usage{
ypr_schedule(population, complete = FALSE)
}
\arguments{
\item{population}{A list of population life-history parameters.}

\item{complete}{A flag indicating whether to generate a complete schedule.}
}
\value{
A tibble of the life-history schedule by age.
}
\description{
Generates a tibble of the life-history schedule by age for a population.
}
\examples{
ypr_schedule(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{plot.ypr_population}}.
}
