% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate-schedule-deprec.R
\name{ypr_schedule}
\alias{ypr_schedule}
\title{Life-History Schedule}
\usage{
ypr_schedule(population)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

DEPRECATED: Replace \code{ypr_schedule()} with \code{ypr_tabulate_schedule()}
}
\details{
Generates the life-history schedule by age for a population.
}
\keyword{internal}
