% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_biomass}
\alias{ypr_tabulate_biomass}
\title{Tablulate Biomass (and Eggs)}
\usage{
ypr_tabulate_biomass(population)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}
}
\value{
A data frame
}
\description{
Produces a data frame of the 'Weight' and 'Fecundity' and the number of
'Survivors' and 'Spawners' and the total 'Biomass' and 'Eggs' by
'Age' class.
}
\examples{
ypr_tabulate_biomass(ypr_population())
}
\seealso{
Other tabulate: 
\code{\link{ypr_tabulate_fish}()},
\code{\link{ypr_tabulate_parameters}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}
}
\concept{tabulate}
