% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_fish}
\alias{ypr_tabulate_fish}
\title{Table Fish}
\usage{
ypr_tabulate_fish(population, x = "Age", binwidth = 1L)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}

\item{x}{A string of the term on the x-axis.}

\item{binwidth}{A positive integer of the width of the bins for grouping.}
}
\value{
A data frame
}
\description{
Produces a data frame of the number of fish in the
'Surviving', 'Spawning', 'Caught', 'Harvested', 'Released' and 'HandlingMortality' categories by
'Length', 'Age' or 'Weight' class.
}
\examples{
ypr_tabulate_fish(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_plot_fish}}
}
