% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate-fish.R
\name{ypr_tabulate_fish}
\alias{ypr_tabulate_fish}
\title{Tabulate Fish Numbers}
\usage{
ypr_tabulate_fish(population, x = "Age", binwidth = 1L)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}

\item{x}{The object to coerce.}

\item{binwidth}{A positive integer of the width of the bins for grouping.}
}
\value{
A data frame
}
\description{
Produces a data frame of the number of fish in the 'Survivors', 'Spawners',
'Caught', 'Harvested', 'Released' and 'HandlingMortalities' categories by
'Length', 'Age' or 'Weight' class and 'Ecotype' (NA if not applicable)
}
\examples{
ypr_tabulate_fish(ypr_population())
}
\seealso{
Other tabulate: 
\code{\link{ypr_detabulate_parameters}()},
\code{\link{ypr_report}()},
\code{\link{ypr_tabulate_biomass}()},
\code{\link{ypr_tabulate_parameters}()},
\code{\link{ypr_tabulate_schedule}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}

Other fish: 
\code{\link{ypr_plot_fish}()}
}
\concept{fish}
\concept{tabulate}
