% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_parameters}
\alias{ypr_tabulate_parameters}
\title{Population Parameters}
\usage{
ypr_tabulate_parameters(population)
}
\arguments{
\item{population}{An object of class \code{\link{ypr_population}}.}
}
\value{
A table of population parameters
}
\description{
Population Parameters
}
\examples{
ypr_tabulate_parameters(ypr_population())
}
\seealso{
\code{\link{ypr_population}}
}
