% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/schedule.R
\name{ypr_schedule}
\alias{ypr_schedule}
\alias{ypr_tabulate_schedule}
\title{Life-History Schedule}
\usage{
ypr_schedule(population)

ypr_tabulate_schedule(population)
}
\arguments{
\item{population}{An object of class \code{\link[=ypr_population]{ypr_population()}}.}
}
\value{
A tibble of the life-history schedule by age.
}
\description{
Generates the life-history schedule by age for a population.
}
\section{Functions}{
\itemize{
\item \code{ypr_schedule}: Tabulate Schedule (Deprecated)
}}

\examples{
ypr_tabulate_schedule(ypr_population())
}
\seealso{
Other tabulate: 
\code{\link{ypr_detabulate_parameters}()},
\code{\link{ypr_report}()},
\code{\link{ypr_tabulate_biomass}()},
\code{\link{ypr_tabulate_fish}()},
\code{\link{ypr_tabulate_parameters}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}

Other schedule: 
\code{\link{ypr_plot_schedule}()}
}
\concept{schedule}
\concept{tabulate}
