% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate-schedule.R
\name{ypr_tabulate_schedule}
\alias{ypr_tabulate_schedule}
\alias{ypr_tabulate_schedule.ypr_population}
\alias{ypr_tabulate_schedule.ypr_ecotypes}
\title{Life-History Schedule}
\usage{
ypr_tabulate_schedule(object, ...)

\method{ypr_tabulate_schedule}{ypr_population}(object, ...)

\method{ypr_tabulate_schedule}{ypr_ecotypes}(object, ...)
}
\arguments{
\item{object}{The population or populations.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the life-history schedule by age.
}
\description{
Generates the life-history schedule by age for a population.
}
\section{Methods (by class)}{
\itemize{
\item \code{ypr_tabulate_schedule(ypr_population)}: Tabulate Schedule

\item \code{ypr_tabulate_schedule(ypr_ecotypes)}: Tabulate Schedule

}}
\examples{
ypr_tabulate_schedule(ypr_population())
ypr_tabulate_schedule(ypr_ecotypes(Linf = c(10, 20)))
}
\seealso{
Other tabulate: 
\code{\link{ypr_detabulate_parameters}()},
\code{\link{ypr_report}()},
\code{\link{ypr_tabulate_biomass}()},
\code{\link{ypr_tabulate_fish}()},
\code{\link{ypr_tabulate_parameters}()},
\code{\link{ypr_tabulate_sr}()},
\code{\link{ypr_tabulate_yields}()},
\code{\link{ypr_tabulate_yield}()}

Other schedule: 
\code{\link{ypr_plot_schedule}()}
}
\concept{schedule}
\concept{tabulate}
