% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_sr}
\alias{ypr_tabulate_sr}
\title{Stock-Recruitment Parameters}
\usage{
ypr_tabulate_sr(object, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{...}{Unused parameters.}
}
\value{
A table of stock-recruitment parameters.
}
\description{
Stock-Recruitment Parameters
}
\examples{
ypr_tabulate_sr(ypr_population()) # Beverton-Holt
ypr_tabulate_sr(ypr_population(BH = 0L)) # Ricker
}
