% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_yield}
\alias{ypr_tabulate_yield}
\title{Tabulate Yield}
\usage{
ypr_tabulate_yield(object, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{...}{Unused parameters.}
}
\value{
A data frame.
}
\description{
Tabulate Yield
}
\examples{
ypr_tabulate_yield(ypr_population())
}
\seealso{
\code{\link{ypr_population}}, \code{\link{ypr_populations}} and \code{\link{ypr_yield}}
}
