% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_yield.ypr_population}
\alias{ypr_tabulate_yield.ypr_population}
\title{Tabulate Yield}
\usage{
\method{ypr_tabulate_yield}{ypr_population}(object, Ly = 0,
  harvest = TRUE, biomass = FALSE, type = "both", all = FALSE, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield for harvested fish or captures.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.}

\item{type}{A string indicating whether to include 'both' or just the 'actual' or 'optimal' yield.}

\item{all}{A flag indicating whether to include all parameter values.}

\item{...}{Unused parameters.}
}
\value{
A data frame.
}
\description{
Tabulate Yield
}
\examples{
ypr_tabulate_yield(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_yield}}
}
