% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{ypr_tabulate_yields}
\alias{ypr_tabulate_yields}
\title{Tabulate Yields}
\usage{
ypr_tabulate_yields(object, ...)
}
\arguments{
\item{object}{The populations or populations.}

\item{...}{Unused parameters.}
}
\value{
A data frame.
}
\description{
Tabulate Yields
}
\examples{
ypr_tabulate_yields(ypr_population())
}
\seealso{
Other tabulate: \code{\link{ypr_detabulate_parameters}},
  \code{\link{ypr_tabulate_biomass}},
  \code{\link{ypr_tabulate_fish}},
  \code{\link{ypr_tabulate_parameters}},
  \code{\link{ypr_tabulate_sr}},
  \code{\link{ypr_tabulate_yield}}
}
\concept{tabulate}
