% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{ypr_yield}
\alias{ypr_yield}
\title{Yield}
\usage{
ypr_yield(x, Ly = 0, harvest = TRUE, biomass = TRUE, sanitize = TRUE,
  check = TRUE)
}
\arguments{
\item{x}{A list of population life-history parameters or a data frame of the complete life-history schedule.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{sanitize}{A flag indicating whether to replace negative or NaNs with 0.}

\item{check}{A flag indicating whether to check the arguments.}
}
\description{
Calculates the yield for a population in terms of the proportion of the unfished population abundance or biomass.
}
\examples{
ypr_yield(ypr_population())
}
\seealso{
\code{\link{ypr_population}}, \code{\link{ypr_optimize}},
\code{\link{ypr_yields}} and \code{\link{ypr_plot}}
}
