% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{ypr_yield}
\alias{ypr_yield}
\title{Yield}
\usage{
ypr_yield(population, Ly = 0, harvest = FALSE, biomass = FALSE,
  sanitize = TRUE, check = TRUE)
}
\arguments{
\item{population}{A list of population life-history parameters (or a data frame of the complete life-history schedule).}

\item{Ly}{The minimum length (trophy) fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over just harvested versus harvested and released fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass (kg) versus number of captured individuals.}

\item{sanitize}{A flag indicating whether to replace negative or NaNs with 0 (internal use only).}

\item{check}{A flag indicating whether to check the arguments (internal use only).}
}
\value{
The yield as number of fish or biomass (kg).
}
\description{
Calculates the yield for a population.
}
\details{
If the yield is given in terms of the number of captured individuals then the scaling depends on the value of \code{R0}.
By default, with \code{R0 = 1} the number is as a proportion of the recruits at the unfished equilibrium.
If the yield is given in terms of the biomass then the scaling depends on the values of \code{Wa} and \code{R0}.
}
\examples{
ypr_yield(ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_optimize}}
}
