% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{ypr_yield}
\alias{ypr_yield}
\title{Yield}
\usage{
ypr_yield(schedule, Rk = 5, Ly = 0, harvest = TRUE, biomass = TRUE)
}
\arguments{
\item{schedule}{A data frame of the life history schedule.}

\item{Rk}{The numbers of spawners per spawner at low density.}

\item{Ly}{The minimum length fish to consider.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}
}
\description{
Yield
}
\examples{
ypr_yield(ypr_schedule())
}
