% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield-uncertainty.R
\name{ypr_yield_uncertainty}
\alias{ypr_yield_uncertainty}
\title{Yield with Uncertainty}
\usage{
ypr_yield_uncertainty(population, population2, Ly = 0, harvest = TRUE,
  biomass = TRUE, n = 10^4, level = 0.95,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{population}{A list of population life-history parameters (or a data frame of the complete life-history schedule).}

\item{population2}{A second list of population life-history parameters.}

\item{Ly}{The minimum length fish to consider when calculating the yield.}

\item{harvest}{A flag indicating whether to calculate the yield over harvested versus captured fish.}

\item{biomass}{A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.}

\item{n}{A numeric of the number of iterations.}

\item{level}{A probability of the interval.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously.
The option is initialized from environment variable MC_CORES if set.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
A double vector of the yield and its lower and upper limits.
}
\description{
Calculates a yield with lower and upper limits.
The uncertainty in each population parameter is assumed to be a uniform distribution between the value in population and the value in population2.
The parameters are assumed to be uncorrelated and there is no requirement
for the value in population to be less than the value in population2
The parameters \code{tmax} and \code{Rk} must not vary.
}
\examples{
population <- ypr_population()
ypr_yield_uncertainty(ypr_population_update(population, Rk = 10), ypr_population())
}
\seealso{
\code{\link{ypr_population}} and \code{\link{ypr_yield}}
}
